move android code
